Refactor x.py and fix benchmarking.
